requirements.txt
streamlit==1.15.0
bcrypt==3.2.0
pymongo==3.12.1
python-dotenv==0.19.2
Pillow==8.4.0
transformers==4.12.3
requests==2.26.0